refactor: move commands modules to separate folder
