+++ src/commands/create.js
@@ -1,5 +1,5 @@
 import fs from "fs";
-import { errorHandler } from "./utils/errorHandler.js";
+import { errorHandler } from "../utils/errorHandler.js";
 
 export const create = async (newFileUrl) => {
   fs.open(newFileUrl, "wx", (err) => {
