+++ src/commands/decompress.js
@@ -1,18 +1,18 @@
 import fs from "fs";
 import zlib from "zlib";
-import { errorHandler } from "./utils/errorHandler.js";
+import { errorHandler } from "../utils/errorHandler.js";
 
-export const compress = (fileToCompressUrl, destinationFileUrl) => {
-  const readable = fs.createReadStream(fileToCompressUrl);
+export const decompress = (fileToDecompressUrl, destinationFileUrl) => {
+  const readable = fs.createReadStream(fileToDecompressUrl);
   const writable = fs.createWriteStream(destinationFileUrl);
 
-  const brotli = zlib.createBrotliCompress();
+  const brotli = zlib.createBrotliDecompress();
 
   readable
     .pipe(brotli)
     .pipe(writable)
     .on("finish", () => {
-      console.log("Successfully done compessing");
+      console.log("Successfully done decompessing");
       writable.end();
     });
