+++ src/commands/getHash.js
@@ -1,7 +1,7 @@
 import { createReadStream } from "node:fs";
 import { stdout } from "node:process";
-import { errorHandler } from "./utils/errorHandler.js";
 const { createHash } = await import("node:crypto");
+import { errorHandler } from "../utils/errorHandler.js";
 
 export const calculateHash = async (fileUrl) => {
   const readable = createReadStream(fileUrl);
