+++ src/commands/getOsData.js
