+++ src/commands/showAvailableCommands.js
